python_ai_server/requirements.txt
fastapi
uvicorn
langchain
openai
httpx
python-dotenv
pydantic
requests